requirements.txt
aiodanbooru==1.1.1
aiohttp==3.9.5
discord.py==2.4.0
Flask==3.0.3
google-generativeai==0.7.2
psutil==6.0.0
protobuf==5.27.3
python-dotenv==1.0.1
qrcode==7.4.2
Requests==2.32.3
scratchattach==1.7.3
tzdata==2024.1
yt_dlp==2023.7.6
